+++ ParseSwift.playground/Pages/5 - ACL.xcplaygroundpage/Contents.swift
@@ -33,7 +33,15 @@ struct GameScore: ParseObject {
     //: Your own properties
     var score: Int
 
-    //: a custom initializer
+    init() {
+        self.score = 0
+    }
+}
+
+//: It's recommended to place custom initializers in an extension
+//: to preserve the convenience initializer.
+extension GameScore {
+    //: Custom initializer.
     init(score: Int) {
         self.score = score
     }